django_example/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.13
django-constance==3.1.0
django-picklefield==3.2
idna==3.7
josepy==1.14.0
mozilla-django-oidc==4.0.1
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
